ofbiz merge: manual updates to e7777d13468 from trunk (minilang component)
Several manual merges to minilang component files (up to given
revision), including:
* mostly small code cleanups

Ofbiz-Version-Commit: e7777d13468236ae3f3f8af43d2fcfa56c019700
Ofbiz-Repo: https://github.com/apache/ofbiz-framework.git
Ofbiz-Branch: trunk

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/556

+++ framework/minilang/src/org/ofbiz/minilang/method/otherops/Trace.java
@@ -32,7 +32,7 @@ import org.w3c.dom.Element;
 /**
  * Implements the &lt;trace&gt; element.
  * 
- * @see <a href="https://cwiki.apache.org/confluence/display/OFBADMIN/Mini+Language+-+minilang+-+simple-method+-+Reference">Mini-language Reference</a>
+ * @see <a href="https://cwiki.apache.org/confluence/display/OFBIZ/Mini+Language+-+minilang+-+simple-method+-+Reference">Mini-language Reference</a>
  */
 public final class Trace extends MethodOperation {
 
@@ -53,7 +53,7 @@ public final class Trace extends MethodOperation {
             MiniLangValidate.handleError("Invalid level attribute", simpleMethod, element);
             this.level = Debug.INFO;
         } else {
-            this.level = levelInt.intValue();
+            this.level = levelInt;
         }
         methodOperations = Collections.unmodifiableList(SimpleMethod.readOperations(element, simpleMethod));
     }
